Move common functions to commons
