+++ commons.py
@@ -10,6 +10,7 @@ else:
     import atexit
     from select import select
 
+from datetime import datetime
 from pathlib import Path
 
 
@@ -99,3 +100,24 @@ class KBHit:
         else:
             dr, dw, de = select([sys.stdin], [], [], 0)
             return dr != []
+
+
+class BaseClass:
+    """
+    Base class for both bank and inspector
+    """
+
+    def _log(self, message, stdio=True, in_file=False, file_mode="a+"):
+        prefix = datetime.now().strftime("%Y-%m-%d:%H:%M:%S ")
+
+        if stdio:
+            print(prefix + str(message))
+
+        if in_file:
+            with open(self.log_database, mode=file_mode) as f:
+                f.write(prefix + str(message) + "\n")
+
+    def _id_to_index(self, bid):
+        for i, branch in enumerate(self.branches):
+            if branch["_id"] == bid:
+                return i
