+++ inspector.py
@@ -1,14 +1,13 @@
 import pickle
 import socket
 import threading
-from datetime import datetime
 from threading import Thread
 
 from bank import Bank
-from commons import Constants
+from commons import BaseClass, Constants
 
 
-class Inspector:
+class Inspector(BaseClass):
     def __init__(self, address="localhost"):
 
         self.address = address
@@ -132,18 +131,3 @@ class Inspector:
             threads[-1].start()
         for i in range(Bank.n_branches):
             threads[i].join()
-
-    def _log(self, message, stdio=True, in_file=False, file_mode="a+"):
-        prefix = datetime.now().strftime("%Y-%m-%d:%H:%M:%S ")
-
-        if stdio:
-            print(prefix + str(message))
-
-        if in_file:
-            with open(self.log_database, mode=file_mode) as f:
-                f.write(prefix + str(message) + "\n")
-
-    def _id_to_index(self, bid):
-        for i, branch in enumerate(self.branches):
-            if branch["_id"] == bid:
-                return i
